feat(admin.playlist): extract components and enhance filter functionality
